native hints refactored
